refactoring inner class to upper
