+++ src/main/java/com/mageddo/commons/exec/CommandLines.java
@@ -1,34 +1,14 @@
 package com.mageddo.commons.exec;
 
-import com.mageddo.io.LogPrinter;
-import com.mageddo.wait.Wait;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.SneakyThrows;
-import lombok.ToString;
-import lombok.Value;
-import lombok.experimental.NonFinal;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.exec.CommandLine;
-import org.apache.commons.exec.DaemonExecutor;
 import org.apache.commons.exec.ExecuteException;
 import org.apache.commons.exec.ExecuteResultHandler;
-import org.apache.commons.exec.ExecuteStreamHandler;
 import org.apache.commons.exec.ExecuteWatchdog;
-import org.apache.commons.exec.Executor;
-import org.apache.commons.exec.PumpStreamHandler;
-import org.apache.commons.lang3.Validate;
 
-import java.io.ByteArrayOutputStream;
-import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.io.UncheckedIOException;
 import java.time.Duration;
-import java.util.Map;
-import java.util.function.Supplier;
 
 @Slf4j
 public class CommandLines {
@@ -106,192 +86,4 @@ public class CommandLines {
     return new ProcessAccessibleDaemonExecutor();
   }
 
-  @Getter
-  static class ProcessAccessibleDaemonExecutor extends DaemonExecutor {
-
-    private Process process = null;
-
-    @Override
-    protected Process launch(CommandLine command, Map<String, String> env, File dir) throws IOException {
-      return this.process = super.launch(command, env, dir);
-    }
-  }
-
-  @Value
-  @Builder(toBuilder = true, builderClassName = "RequestBuilder", buildMethodName = "build0")
-  public static class Request {
-
-    @NonNull
-    private final CommandLine commandLine;
-
-    private final Duration timeout;
-
-    private final ExecuteResultHandler handler;
-    private Map<String, String> env;
-
-    @NonFinal
-    private boolean watchingOutput;
-
-    @Builder.Default
-    private final Streams streams = Streams.builder()
-      .outAndErr(new ByteArrayOutputStream())
-      .build();
-
-    public ExecuteStreamHandler getStreamHandler() {
-      return this.streams.toStreamHandler();
-    }
-
-    private Request printOutToLogsInBackground() {
-      if (this.watchingOutput) {
-        throw new IllegalStateException("Already watching output");
-      }
-      this.watchingOutput = true;
-      LogPrinter.printInBackground(this.streams.outAndErr.getPipedIn());
-      return this;
-    }
-
-    public OutputStream getBestOut() {
-      return this.streams.getBestOriginalOutput();
-    }
-
-    public long getTimeoutInMillis() {
-      if (this.timeout == null) {
-        return ExecuteWatchdog.INFINITE_TIMEOUT;
-      }
-      return this.timeout.toMillis();
-    }
-
-    public static class RequestBuilder {
-
-      private boolean printLogsInBackground = false;
-
-      public Request build() {
-        final var request = this.build0();
-        if (this.printLogsInBackground) {
-          request.printOutToLogsInBackground();
-        }
-        return request;
-      }
-
-      public RequestBuilder printLogsInBackground() {
-        this.printLogsInBackground = true;
-        return this;
-      }
-    }
-
-
-    @Value
-    @Builder(toBuilder = true, builderClassName = "StreamsBuilder")
-    public static class Streams {
-
-      private final PipedStream outAndErr;
-      private final OutputStream out;
-      private final OutputStream err;
-      private final InputStream input;
-
-      public PipedStream getBestOut() {
-        if (this.outAndErr != null) {
-          return this.outAndErr;
-        }
-        throw new UnsupportedOperationException();
-      }
-
-      public OutputStream getBestOriginalOutput() {
-        return this.getBestOut().getOriginalOut();
-      }
-
-      public static class StreamsBuilder {
-        public StreamsBuilder outAndErr(OutputStream outAndErr) {
-          this.outAndErr = new PipedStream(outAndErr);
-          return this;
-        }
-      }
-
-      public ExecuteStreamHandler toStreamHandler() {
-        if (this.outAndErr != null) {
-          return new PumpStreamHandler(this.outAndErr);
-        }
-        return new PumpStreamHandler(this.out, this.err, this.input);
-      }
-    }
-  }
-
-  @Getter
-  @Builder
-  @ToString(of = {"exitCode"})
-  public static class Result {
-
-    @NonNull
-    private Request request;
-
-    @NonNull
-    private Executor executor;
-
-    @NonNull
-    private OutputStream out;
-
-    @NonNull
-    private Supplier<Process> processSupplier;
-
-    private Integer exitCode;
-
-    public Result printOutToLogsInBackground() {
-      this.request.printOutToLogsInBackground();
-      return this;
-    }
-
-    public String getOutAsString() {
-      Validate.isTrue(this.out instanceof ByteArrayOutputStream, "Only ByteArrayOutputStream is supported");
-      return this.out.toString();
-    }
-
-    public Result checkExecution() {
-      if (this.executor.isFailure(this.getExitCode())) {
-        throw new ExecutionValidationFailedException(this);
-      }
-      return this;
-    }
-
-    public String toString(boolean printOut) {
-      return String.format(
-        "code=%d, out=%s",
-        this.exitCode, printOut ? this.getOutAsString() : null
-      );
-    }
-
-    @SneakyThrows
-    public Process getProcess() {
-      return this.processSupplier.get();
-    }
-
-    public Long getProcessId() {
-      final var process = this.getProcess();
-      if (process == null) {
-        return null;
-      }
-      return process.pid();
-    }
-
-    public void waitProcessToFinish() {
-      new Wait<>()
-        .infinityTimeout()
-        .ignoreException(IllegalArgumentException.class)
-        .until(() -> {
-          Validate.isTrue(this.isProcessFinished(), "Process not finished yet");
-          return true;
-        });
-    }
-
-    private boolean isProcessFinished() {
-      return getProcess() != null && !getProcess().isAlive();
-    }
-
-    public Integer getProcessExitCodeWhenAvailable() {
-      try {
-        return getProcess().exitValue();
-      } catch (IllegalThreadStateException e) {
-        return null;
-      }
-    }
-  }
 }
