+++ src/main/java/com/mageddo/dnsproxyserver/sandbox/BinaryFromGradleTestsSandbox.java
@@ -2,6 +2,7 @@ package com.mageddo.dnsproxyserver.sandbox;
 
 import com.mageddo.commons.exec.CommandLines;
 import com.mageddo.commons.exec.NopResultHandler;
+import com.mageddo.commons.exec.Request;
 import com.mageddo.dnsproxyserver.config.dataprovider.vo.ConfigEnv;
 import lombok.extern.slf4j.Slf4j;
 
@@ -12,7 +13,7 @@ import java.util.Map;
 public class BinaryFromGradleTestsSandbox {
   public Instance run(Path configFile) {
     final var commandLine = DpsBinaryExecutableFinder.buildCommandLine();
-    final var request = CommandLines.Request.builder()
+    final var request = Request.builder()
       .commandLine(commandLine)
       .handler(new NopResultHandler())
       .env(Map.of(ConfigEnv.MG_CONFIG_FILE_PATH, configFile.toString()))
